trying to improve compression ratio

+++ src/utils/SlidingWindow.cpp
@@ -14,7 +14,7 @@ const array<_type, _size> &SlidingWindow<_type, _size>::getWindow() const {
 }
 
 template<typename _type, size_t _size>
-const _type &SlidingWindow<_type, _size>::readAtIndex(const size_t &index) const {
+const _type &SlidingWindow<_type, _size>::readAtIndex(const int &index) const {
     return mWindow[getStandardEquivalentIndex(index)];
 }
 
@@ -25,7 +25,7 @@ void SlidingWindow<_type, _size>::insert(const _type &newValue) {
 }
 
 template<typename _type, size_t _size>
-size_t SlidingWindow<_type, _size>::getStandardEquivalentIndex(const size_t &index) const {
+size_t SlidingWindow<_type, _size>::getStandardEquivalentIndex(const int &index) const {
     size_t standardIndex = index % _size;
     if (standardIndex < 0) {
         standardIndex += _size;
